requirements.txt
aiogram==3.12.0
beautifulsoup4==4.12.3
langchain==0.2.15
langchain_chroma==0.1.3
langchain_community==0.2.15
langchain_core==0.2.37
langchain_openai==0.1.23
langchain_text_splitters==0.2.2
openai==1.43.0
pinecone==5.1.0
pinecone_text==0.9.0
psycopg2==2.9.9
python-dotenv==1.0.1
python-telegram-bot==21.5
Requests==2.32.3
tiktoken==0.7.0
python_telegram_bot==21.5
